Refactor NodeFactory to use separate module for node types
